pyproject.toml
[project]
name = "text2ink"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fitz>=0.0.1.dev2",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pymupdf>=1.26.5",
]
